+++ packages/clients/vue/src/components/flow/index.ts
@@ -54,6 +54,8 @@ export type { FlowContextValue, FlowProps } from "./Flow.vue";
 // Root component
 export { default as Flow } from "./Flow.vue";
 export { default as FlowCancel } from "./FlowCancel.vue";
+export { default as FlowPause } from "./FlowPause.vue";
+export { default as FlowResume } from "./FlowResume.vue";
 export type {
   FlowDropZoneProps,
   FlowDropZoneSlotProps,