+++ packages/flow/utility/nodes/src/nodes/index.ts
@@ -1,4 +1,5 @@
 export * from "./conditional-node";
 export * from "./merge-node";
 export * from "./multiplex-node";
+export * from "./passthrough-node";
 export * from "./zip-node";